+++ src/pages/main-page/ui/main-page-cities/index.tsx
@@ -1,0 +1,3 @@
+import { MainPageCities } from './main-page-cities';
+
+export { MainPageCities };
