+++ src/pages/main-page/ui/main-page/main-page.tsx
@@ -1,10 +1,7 @@
 import Layout from '@widgets/layout';
-import type { MainOfferInfo, OfferCityName } from '@entities/offer';
+import type { MainOfferInfo } from '@entities/offer';
 import { componentWithBrowserTitle } from '@shared/hoc/component-with-browser-title';
-import OffersList from '@features/offers-list';
-import NoPlacesSection from '../no-places-section';
-import { ComponentProps, useEffect, useMemo, useState } from 'react';
-import { Nullable } from '@shared/types';
+import { useEffect, useMemo } from 'react';
 import LocationsFilterList, { useActiveLocation } from '@features/locations-filter-list/';
 import { ALL_CITIES_NAMES, PAGE_TITLE, DEFAULT_CITY, DEFAULT_SEARCH_PARAMS } from '@pages/main-page/config';
 import { useSearchParams } from 'react-router-dom';
@@ -12,36 +9,14 @@ import { getSearchParam } from '@shared/lib/url';
 import { SearchParams } from '@pages/main-page/model';
 import { getPageStyles } from './get-styles';
 import { isOfferCityName } from './type-guards';
-import LeafletMap from '@features/leaflet-map';
-import { OfferSortingSelect, offerSortTypeToComparerMap, useOfferSorting } from '@features/offer-sorting-select';
+import { offerSortTypeToComparerMap, useOfferSorting } from '@features/offer-sorting-select';
+import { MainPageCities } from '../main-page-cities';
 
 type MainPageProps = {
   offers: MainOfferInfo[];
 }
 
-function getLeafletMapProps(offers: MainOfferInfo[], activeOfferId: Nullable<string>): ComponentProps<typeof LeafletMap> {
-  const activeOffer = activeOfferId
-    ? offers.find((current) => current.id === activeOfferId)
-    : undefined;
-
-  return {
-    center: {
-      location: offers[0].city.location,
-      name: offers[0].city.name
-    },
-    points: offers.map((current) => ({
-      location: current.location,
-      name: current.title
-    })),
-    selectedPoint: activeOffer ? {
-      location: activeOffer.location,
-      name: activeOffer.title
-    } : null
-  };
-}
-
 function MainPage({ offers }: MainPageProps): JSX.Element {
-  const [activeOfferId, setActiveOfferId] = useState<Nullable<string>>(null);
   const [searchParams, setSearchParams] = useSearchParams(DEFAULT_SEARCH_PARAMS);
   const activeCitySearchParam = getSearchParam<SearchParams, keyof SearchParams>(searchParams, 'activeCity', DEFAULT_CITY);
   const { activeLocation } = useActiveLocation(isOfferCityName(activeCitySearchParam) ? activeCitySearchParam : DEFAULT_CITY);
@@ -50,10 +25,6 @@ function MainPage({ offers }: MainPageProps): JSX.Element {
   const filteredOffers = useMemo(
     () => offers.filter((current) => current.city.name === activeLocation).sort(offerSortTypeToComparerMap.get(activeSotingType)),
     [activeLocation, activeSotingType, offers]
-  );
-  const mapProps = useMemo(
-    () => filteredOffers.length ? getLeafletMapProps(filteredOffers, activeOfferId) : null,
-    [filteredOffers, activeOfferId]
   );
 
   const isOffersExists = filteredOffers.length > 0;
@@ -72,12 +43,6 @@ function MainPage({ offers }: MainPageProps): JSX.Element {
 
   const { containerClassName, contentClassName } = getPageStyles(isOffersExists);
 
-  const activeCityChangeHandler = (cityName: OfferCityName) => {
-    if (activeLocation !== cityName) {
-      setActiveOfferId(null);
-    }
-  };
-
   return (
     <Layout className='page--gray page--main'>
       <Layout.Header />
@@ -88,33 +53,15 @@ function MainPage({ offers }: MainPageProps): JSX.Element {
             <LocationsFilterList
               activeFilter={activeLocation}
               allFilterItems={ALL_CITIES_NAMES}
-              onFilterChange={activeCityChangeHandler}
             />
           </section>
         </div>
-        <div className='cities'>
-          <div className={containerClassName}>
-            {isOffersExists ?
-              <section className='cities__places places'>
-                <h2 className='visually-hidden'>Places</h2>
-                <b className='places__found'>{filteredOffers.length} places to stay in {activeLocation}</b>
-                <form className='places__sorting' action='#' method='get'>
-                  <span className='places__sorting-caption'>Sort by&nbsp;</span>
-                  <OfferSortingSelect />
-                </form>
-                <OffersList offers={filteredOffers} onActivateOffer={setActiveOfferId} className='tabs__content' />
-              </section>
-              :
-              <NoPlacesSection />}
-            <div className='cities__right-section'>
-              {(isOffersExists && mapProps) &&
-                <LeafletMap
-                  className='cities__map'
-                  {...mapProps}
-                />}
-            </div>
-          </div>
-        </div>
+        {isOffersExists &&
+          <MainPageCities
+            activeLocation={activeLocation}
+            offers={filteredOffers}
+            className={containerClassName}
+          />}
       </Layout.Content>
     </Layout>
   );
