+++ src/widgets/layout/ui/logo/index.ts
@@ -1,0 +1,3 @@
+import { Logo } from './logo';
+
+export { Logo };
